Move v2 presenter infrastructure to the sarcophagus
Authored-by: Greg Cobb <gcobb@pivotal.io>
