+++ vcap_sarcophagus/cloud_controller/rest_controller/preloaded_object_serializer.rb
@@ -1,4 +1,4 @@
-require 'presenters/v2/presenter_provider'
+require 'presenters/v2/presenters'
 
 module VCAP::CloudController::RestController
   class PreloadedObjectSerializer
