+++ vcap_sarcophagus/cloud_controller/rest_controller/presenters/base_presenter.rb
