+++ vcap_sarcophagus/cloud_controller/rest_controller/presenters/default_presenter.rb
